+++ .circleci/deploy-if-version-bump.sh
@@ -5,7 +5,7 @@ if ! git rev-parse `python setup.py --version` 2>/dev/null ; then
     git tag `python setup.py --version`
     git push --tags  # update the repository version
     python setup.py bdist_wheel  # build this package in the dist directory
-    twine upload dist/* --username $PYPI_USERNAME --password $PYPI_PASSWORD  # publish
+    twine upload dist/* --username $PYPI_USERNAME --password $PYPI_PASSWORD  # publish, these environment variables are set within the circleci interface
 else
     echo "Only non-functional elements were modified in this change, there is no need to deploy."
     echo